+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🔧 Fixed
+
+* Emit warnings as `UserWarning` to make sure they are recorded in the logs.
+
+
 ## 0.3.4 - 2024-11-12
 
 ### 🏷️ Changed